fix(pipeline-job): 최신 job 조회 요청 검증을 보장
